requirements.txt
adbutils==0.8.2
allure-pytest==2.9.45
allure-python-commons==2.9.45
apkutils2==1.0.0
Appium-Python-Client==1.0.2
APScheduler==3.6.3
attrdict==2.0.1
attrs==20.3.0
beautifulsoup4==4.10.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
chardet==3.0.4
cigam==0.0.3
decorator==4.4.2
Deprecated==1.2.10
deprecation==2.1.0
docopt==0.6.2
et-xmlfile==1.0.1
execnet==1.9.0
facebook-wda==1.2.0
google-api-core==2.1.1
google-api-python-client==2.27.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==2.10
iniconfig==1.1.1
jdcal==1.4.1
Js2Py==0.71
logzero==1.6.2
lxml==4.6.1
multi-key-dict==2.0.3
mysql-connector==2.2.9
numpy==1.19.5
openpyxl==3.0.6
packaging==20.4
pandas==1.2.1
pbr==5.5.1
Pillow==8.0.1
pipwin==0.5.1
pluggy==0.13.1
pocketsphinx==0.1.15
progress==1.5
protobuf==3.18.1
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyelftools==0.27
pyjsparser==2.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytest==6.1.2
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-warnings==0.3.1
pytest-xdist==2.4.0
python-dateutil==2.8.1
python-jenkins==1.7.0
pytz==2020.4
PyYAML==6.0
qqbot==2.3.11
removebg==0.4
requests==2.24.0
retry==0.9.2
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
uiautomator2==2.12.0
uritemplate==4.1.1
urllib3==1.25.11
weditor==0.6.1
whichcraft==0.6.1
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
